memcube-political/requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0
requests>=2.25.0
jsonlines>=3.0.0
python-dotenv>=0.19.0
loguru>=0.6.0
networkx>=2.6.0

# 图数据库依赖 (可选)
neo4j>=5.0.0
python-arango>=7.0.0

# 向量数据库依赖 (可选)
chromadb>=0.4.0
qdrant-client>=1.6.0
faiss-cpu>=1.7.0

# 其他工具
dataclasses-json>=0.5.0
typing-extensions>=4.0.0